Move user models from 'files' to 'users'
